[Feat]: Stamina, Satiety, Health UI logic

+++ TPS/Assets/Scripts/Character States/ScriptableObject/PlayerDate_SO.cs
@@ -7,8 +7,8 @@ public class PlayerDate_SO : ScriptableObject
 {
     [Header("State Info")]
     public int maxHealth;
-    public int maxStamina; // 体力
-    public int maxSatiety; // 饱食度
-    public int maxInfectivity; // 感染性
+    public float maxStamina; // 体力
+    public float maxSatiety; // 饱食度
+    public float maxInfectivity; // 感染性
     public bool isAlive;
 }
